Attempted fix of handranking, may  be issue in resolution. Can leave ended game. Added more checks to try to stop all ind or folded or finished players from having a turn.

+++ api/src/main/java/com/pokerface/pokerapi/game/GameController.java
@@ -36,7 +36,8 @@ public class GameController {
     private final UserService userService;
     private final AIService aiService;
     private final SimpMessagingTemplate messenger;
-    private List gameHashList = new ArrayList();
+    private boolean ping=true;
+
 
     public GameController(final GameService gameService,
                           final UserService userService,
@@ -54,12 +55,21 @@ public class GameController {
      */
     @MessageMapping("/game/{game_id}/control")
     public void receiveManagementAction() {
-
     }
 
-    @Scheduled(fixedRate = 1000)
+//    @Scheduled(fixedDelay = 10000)
+//    public void checkUp(){
+//        if (ping){
+//            ping=false;
+//        } else {
+//            logger.debug("checkedEventsStopped?");
+//        }
+//    }
+
+    @Scheduled(fixedDelay = 1000)
     public void checkEvents() {
         try {
+            ping=true;
             List<Long> startedGames = gameService.startingGameIDs();
             for (Long gameID : startedGames) {
                 GameStateTransport gameTransport = gameService.gameStart(gameID);
@@ -84,9 +94,10 @@ public class GameController {
         }
     }
 
-    @Scheduled(fixedRate=5000)
+    @Scheduled(fixedDelay=5000)
     public void AILivelinessFix(){
         List<Long> gamesToCheck=gameService.getPotentialAIGames();
+        Iterable gamesTest = gameService.getGameList();
         for (Long gameID:gamesToCheck){
             GameStateTransport nextGameState=gameService.getGameStateTransport(gameID);
                 GameAction aiAction = aiService.playAction(gameService,gameID);
@@ -288,13 +299,15 @@ public class GameController {
         if (user.getId() != userID) {
             throw new UserAccessNotPermittedException(user.getUsername(), gameID, "deletion of user " + userID);
         }
-        GameStateTransport gameStateTransport = gameService.getGameStateTransport(gameID);
-        if (gameService.removePlayer(gameID, userService.getUserByUsername(principal.getName()).getId())==1){
-            closeGame(gameID);
-        } else {
+        if (gameService.doesGameExist(gameID)) {
+            GameStateTransport gameStateTransport = gameService.getGameStateTransport(gameID);
+            if (gameService.removePlayer(gameID, userService.getUserByUsername(principal.getName()).getId()) == 1) {
+                closeGame(gameID);
+            } else {
 
-            gameStateTransport.reason(GameStateTransport.Reason.PLAYER_LEFT, principal.getName() + " has left.");
-            messenger.convertAndSend("/messages/games/" + gameID, gameStateTransport);
+                gameStateTransport.reason(GameStateTransport.Reason.PLAYER_LEFT, principal.getName() + " has left.");
+                messenger.convertAndSend("/messages/games/" + gameID, gameStateTransport);
+            }
         }
 
         return new ResponseEntity(HttpStatus.NO_CONTENT);
